Controler Test Vocacional 2/3, falta correcion de errores

+++ Kihon.sql
@@ -86,10 +86,10 @@ foreign key (mbti) references personalidad(mbti) on delete cascade on update cas
 
 create table test_voc(
 id_res int auto_increment primary key not null,
-ICFM int,
-CSA int,
-CMB int,
-HA int,
+ICFM float,
+CSA float,
+CMB float,
+HA float,
 usuario varchar(20),
 foreign key (usuario) references personal(trabajador) on delete cascade on update cascade
 );
